+++ liblangutil/EVMVersion.cpp
@@ -51,6 +51,9 @@ bool EVMVersion::hasOpcode(Instruction _opcode) const
 		return hasBaseFee();
 	case Instruction::BLOBBASEFEE:
 		return hasBlobBaseFee();
+	case Instruction::TSTORE:
+	case Instruction::TLOAD:
+		return supportsTransientStorage();
 	default:
 		return true;
 	}