+++ libyul/SideEffects.h
@@ -76,11 +76,14 @@ struct SideEffects
 	/// or `None` respectively. Note that, when the value is `Read`, the expression can have an
 	/// effect on `msize()`.
 	Effect memory = None;
+	/// Can write, read or have no effect on transient storage, when the value of `transientStorage` is `Write`, `Read`
+	/// or `None` respectively.
+	Effect transientStorage = None;
 
 	/// @returns the worst-case side effects.
 	static SideEffects worst()
 	{
-		return SideEffects{false, false, false, false, false, Write, Write, Write};
+		return SideEffects{false, false, false, false, false, Write, Write, Write, Write};
 	}
 
 	/// @returns the combined side effects of two pieces of code.
@@ -94,7 +97,8 @@ struct SideEffects
 			cannotLoop && _other.cannotLoop,
 			otherState + _other.otherState,
 			storage + _other.storage,
-			memory + _other.memory
+			memory + _other.memory,
+			transientStorage + _other.transientStorage
 		};
 	}
 
@@ -115,7 +119,8 @@ struct SideEffects
 			cannotLoop == _other.cannotLoop &&
 			otherState == _other.otherState &&
 			storage == _other.storage &&
-			memory == _other.memory;
+			memory == _other.memory &&
+			transientStorage == _other.transientStorage;
 	}
 };